+++ src/Player/ChannelData.h
@@ -1,0 +1,12 @@
+//
+//  ChannelData.h
+//  Tracktion
+//
+//  Created by Brian Arne on 8/25/23.
+//
+
+struct ChannelData
+{
+    int sample; // Replace with juce::AudioBuffer<T>;
+    int effect; // Replace with Effect class
+};
