+++ src/Player/Pattern.cpp
@@ -5,16 +5,15 @@
 //  Created by Brian Arne on 6/2/23.
 //
 
-#include "Pattern.hpp"
-#include <vector>
+#include "Pattern.h"
 
-struct Pattern {
-    int lines;
-    int channels;
-    
-    int effects;
-    
-    
-    std::vector<int> data;
-};
 
+Pattern::Pattern()
+{
+    
+}
+
+Pattern::~Pattern()
+{
+    
+}
